# physicalplan: plan ordered aggregations