--- a/dynparquet/schema.go
+++ b/dynparquet/schema.go
@@ -569,6 +569,14 @@ func (s *Schema) Columns() []ColumnDefinition {
 	return s.columns
 }
 
+func (s *Schema) SortingColumns() []ColumnDefinition {
+	sCols := make([]ColumnDefinition, len(s.sortingColumns))
+	for i, col := range s.sortingColumns {
+		sCols[i] = s.columns[s.columnIndexes[col.ColumnName()]]
+	}
+	return sCols
+}
+
 func (s *Schema) ParquetSchema() *parquet.Schema {
 	switch schema := s.def.(type) {
 	case *schemav2pb.Schema: