--- a/logictest/testdata/exec/aggregate/ordered_aggregate
+++ b/logictest/testdata/exec/aggregate/ordered_aggregate
@@ -1,0 +1,44 @@
+createtable schema=default
+----
+
+insert cols=(example_type, labels.label1, timestamp, value)
+type1   value1  1   1
+type1   value1  2   1
+type2   value1  1   1
+----
+
+insert cols=(example_type, labels.label1, timestamp, value)
+type1   value1  1   1
+type1   value1  2   1
+type2   value1  1   1
+----
+
+insert cols=(example_type, labels.label2, timestamp, value)
+type1   value2  1   1
+type1   value2  2   1
+type2   value2  1   1
+----
+
+insert cols=(example_type, labels.label3, timestamp, value)
+type1   value3  1   1
+type1   value3  2   1
+type2   value3  1   1
+----
+
+exec
+select sum(value) as value_sum group by (example_type, labels.label1)
+----
+type1   null    4
+type2   null    2
+type1   value1  4
+type2   value1  2
+
+exec
+select sum(value) as value_sum group by (example_type, labels)
+----
+type1   value1  4
+type1   value2  2
+type1   value3  2
+type2   value1  2
+type2   value2  1
+type2   value3  1