--- a/query/physicalplan/planordering.go
+++ b/query/physicalplan/planordering.go
@@ -1,0 +1,1 @@
+package physicalplan