This PR is a bunch of commits with fixes and planning changes in order to plan ordered aggregations when the grouping columns are a prefix of ordering columns. This includes cases in which a filter "covers" a prefix of the ordering columns up to the group columns (e.g. when the order is "name, labels, timestamp" and we filter name="somename" and group by (labels,timestamp)). For more explicit information on what cases are and aren't covered, please refer to the `plan/aggregate` logic test which goes into more detail.

Ordered aggregations are disabled by default and can be enabled (as is done with logic tests) by calling `query.NewEngine` and adding the `WithOrderedAggregations` option given that there are most likely some latent bugs.

The performance improvement over hash aggregates is reasonable as expected and I believe that there is some more room for improvement. I want to write up a couple of issues with some ideas for perf improvements on the current state:
```
name           old time/op    new time/op    delta
Query/Range-8    4.82ms ± 1%    4.51ms ± 0%   -6.47%  (p=0.000 n=10+8)

name           old alloc/op   new alloc/op   delta
Query/Range-8    22.3MB ± 0%    21.4MB ± 0%   -4.32%  (p=0.000 n=10+10)

name           old allocs/op  new allocs/op  delta
Query/Range-8     9.55k ± 0%     7.58k ± 0%  -20.61%  (p=0.000 n=9+10)
```